pyproject.toml
[tool.poetry]
name = "intelligent_rh"
version = "0.1.0"
description = ""
authors = ["AymanElMikh <aymanmikh7@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.115.8"
openai = "^1.61.0"
uvicorn = "^0.34.0"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
start = "uvicorn main:app --reload"